I change the way the package-level documentation is handled.
Created a function rpsftm-package.R that creates the package stuff.

The package is the same name as the key function. This remains
?rpsftm give the function
package?rpsftm gives the package
clicking in the package window gives the index

The two could be put together by changing the last line of rpsftm-package.R
to
@doctype package
@name rpfstm