fixed the problem with testthat, and defautl formula.
Need to explicitly set environment of the formula if it is default to be parent.frame().

Also to create the survfit object must tweak the update formula to use the calling environment.

return a value of $fit to allow the update() method to work with changing from the  default formula.