added a basic vignetter directory. Tidied up the immdef data and help fil.
The problem with the comit showing the entire C drive is back.mmm.

I changed the source directory in the build options, as it was refusing to build.
That solved the buildign problem, but _may_ have created the comite problem agiain.

I can't remmber how I fixed it before..grrr